refactor membrane orientation
